requirements.txt
grpcio==1.12.0
torch==0.4.0
